+++ lib/supply_chain.dart
@@ -15,6 +15,7 @@ export 'src/node_blue_print.dart';
 export 'src/node.dart';
 export 'src/plugin.dart';
 export 'src/plugin_blue_print.dart';
+export 'src/plugin_node_replacer.dart';
 export 'src/priority.dart';
 export 'src/producer.dart';
 export 'src/scm.dart';